+++ dbt_coverage/__init__.py
@@ -571,11 +571,16 @@ def check_manifest_version(manifest_json):
         )
 
 
-def load_catalog(project_dir: Path) -> Catalog:
-    catalog_path = project_dir / 'target/catalog.json'
+def load_catalog(project_dir: Path, run_artifacts_dir: Path) -> Catalog:
+    if run_artifacts_dir is None:
+        catalog_path = project_dir / 'target/catalog.json'
+    else:
+        catalog_path = run_artifacts_dir / 'catalog.json'
+
     if not catalog_path.exists():
-        raise FileNotFoundError("target/catalog.json not found - "
+        raise FileNotFoundError("catalog.json not found in target/ or in custom path - "
                                 "before using dbt-coverage, run: dbt docs generate")
+
     with open(catalog_path) as f:
         catalog_json = json.load(f)
 
@@ -587,12 +592,17 @@ def load_catalog(project_dir: Path) -> Catalog:
     return catalog
 
 
-def load_manifest(project_dir: Path) -> Manifest:
-    manifest_path = project_dir / 'target/manifest.json'
+def load_manifest(project_dir: Path, run_artifacts_dir: Path) -> Manifest:
+    if run_artifacts_dir is None:
+        manifest_path = project_dir / 'target/manifest.json'
+    else:
+        manifest_path = run_artifacts_dir / 'manifest.json'
+
     if not manifest_path.exists():
-        raise FileNotFoundError(f"target/manifest.json not found - "
-                                 "before using dbt-coverage, run a dbt command that creates manifest artifact "
-                                 "(see: https://docs.getdbt.com/reference/artifacts/manifest-json)")
+        raise FileNotFoundError(f"manifest.json not found in target or in custom path - "
+                                "before using dbt-coverage, run a dbt command that creates manifest artifact "
+                                "(see: https://docs.getdbt.com/reference/artifacts/manifest-json)")
+
     with open(manifest_path) as f:
         manifest_json = json.load(f)
 
@@ -604,11 +614,14 @@ def load_manifest(project_dir: Path) -> Manifest:
     return manifest
 
 
-def load_files(project_dir: Path) -> Catalog:
-    logging.info("Loading catalog and manifest files from project dir: %s", project_dir)
+def load_files(project_dir: Path, run_artifacts_dir: Path) -> Catalog:
+    if run_artifacts_dir is None:
+        logging.info("Loading catalog and manifest files from project dir: %s", project_dir)
+    else:
+        logging.info("Loading catalog and manifest files from custom dir: %s", run_artifacts_dir)
 
-    catalog = load_catalog(project_dir)
-    manifest = load_manifest(project_dir)
+    catalog = load_catalog(project_dir, run_artifacts_dir)
+    manifest = load_manifest(project_dir, run_artifacts_dir)
 
     for table_name in catalog.tables:
         catalog_table = catalog.get_table(table_name)
@@ -682,16 +695,17 @@ def fail_compare(coverage_report: CoverageReport, compare_path: Path):
                            f"{diff.after.coverage:.2%}")
 
 
-def do_compute(project_dir: Path = Path('.'), cov_report: Path = Path('coverage.json'),
-               cov_type: CoverageType = CoverageType.DOC, cov_fail_under: float = None,
-               cov_fail_compare: Path = None, model_path_filter: Optional[List[str]] = None):
+def do_compute(project_dir: Path = Path('.'), run_artifacts_dir: Path = None,
+               cov_report: Path = Path('coverage.json'), cov_type: CoverageType = CoverageType.DOC,
+               cov_fail_under: float = None, cov_fail_compare: Path = None,
+               model_path_filter: Optional[List[str]] = None):
     """
     Computes coverage for a dbt project.
 
     Use this method in your Python code to bypass typer.
     """
 
-    catalog = load_files(project_dir)
+    catalog = load_files(project_dir, run_artifacts_dir)
 
     if len(model_path_filter) >= 1:
         catalog = catalog.filter_catalog(model_path_filter)
@@ -728,6 +742,8 @@ def do_compare(report: Path, compare_report: Path):
 
 @app.command()
 def compute(project_dir: Path = typer.Option('.', help="dbt project directory path."),
+            run_artifacts_dir: Path = typer.Option(None, help="custom directory path for "
+                                                              "catalog and manifest files"),
             cov_report: Path = typer.Option('coverage.json', help="Output coverage report path."),
             cov_type: CoverageType = typer.Argument(..., help="Type of coverage to compute."),
             cov_fail_under: float = typer.Option(None, help="Fail if coverage is lower than "
@@ -743,8 +759,8 @@ def compute(project_dir: Path = typer.Option('.', help="dbt project directory pa
                                                                              "on.")):
     """Compute coverage for project in PROJECT_DIR from catalog.json and manifest.json."""
 
-    return do_compute(project_dir, cov_report, cov_type, cov_fail_under, cov_fail_compare,
-                      model_path_filter)
+    return do_compute(project_dir, run_artifacts_dir, cov_report, cov_type, cov_fail_under,
+                      cov_fail_compare, model_path_filter)
 
 
 @app.command()